function/requirements.txt
azure-functions
azure-identity
azure-mgmt-advisor
requests
python-dotenv